+++ tests/exp_transformer_test.py
@@ -2,40 +2,36 @@ import jax
 import jax.numpy as jnp
 import haiku as hk
 
-from e3nn_jax import Irreps, radius_graph, spherical_harmonics, sus
+from e3nn_jax import IrrepsData, radius_graph, spherical_harmonics, sus, soft_one_hot_linspace
 from e3nn_jax.experimental.transformer import Transformer
 
 
 def test_transformer(keys):
-    irreps_node_input = Irreps("2x0e + 2x1e + 2x2e")
-    irreps_node_output = Irreps("0e + 2x1e + 2e")
-    irreps_edge_attr = Irreps("0e + 1e + 2e")
-
     @hk.without_apply_rng
     @hk.transform
-    def c(pos, features):
-        src, dst = radius_graph(pos, 2.0, size=10)  # TODO check what happens with -1 indices
-        edge_attr = spherical_harmonics(irreps_edge_attr, pos[dst] - pos[src], True)
-        edge_weight_cutoff = sus(3.0 * (2.0 - jnp.linalg.norm(pos[dst] - pos[src], axis=-1)))
+    def model(pos, src, dst, node_feat):
+        edge_attr = spherical_harmonics("0e + 1e + 2e", pos[dst] - pos[src], True)
+        edge_distance = jnp.linalg.norm(pos[dst] - pos[src], axis=-1)
+        edge_weight_cutoff = sus(3.0 * (2.0 - edge_distance))
+        edge_scalar_attr = soft_one_hot_linspace(edge_distance, 0.0, 2.0, 5, 'smooth_finite', cutoff=True)
 
         return Transformer(
-            irreps_node_input=irreps_node_input,
-            irreps_node_output=irreps_node_output,
-            irreps_edge_attr=irreps_edge_attr,
+            "0e + 2x1e + 2e",
             list_neurons=[32, 32],
-            phi=jax.nn.relu,
+            act=jax.nn.relu,
             num_heads=2,
-        )(src, dst, jnp.ones((src.shape[0], 1)), edge_attr, edge_weight_cutoff, features)
+        )(src, dst, edge_scalar_attr, edge_weight_cutoff, edge_attr, node_feat)
 
-    f = jax.jit(c.apply)
+    apply = jax.jit(model.apply)
 
     pos = jnp.array([
         [0.0, 0.0, 0.0],
         [1.0, 0.0, 0.0],
     ])
-    x = irreps_node_input.randn(next(keys), (pos.shape[0], -1))
+    src, dst = radius_graph(pos, 2.0)
+    node_feat = IrrepsData.randn("2x0e + 2x1e + 2x2e", next(keys), (pos.shape[0],))
 
-    w = c.init(next(keys), pos, x)
-    f(w, pos, x)
+    w = model.init(next(keys), pos, src, dst, node_feat)
+    apply(w, pos, src, dst, node_feat)
 
     # TODO test equivariance